Rename pyfuckery module to fuckery.
