+++ src/fuckery/__init__.py
