+++ src/fuckery/cli.py
