+++ src/fuckery/constants.py
