+++ src/fuckery/exc.py
